requirements.txt
streamlit
python-dotenv
langchain-google-genai
scikit-learn
numpy

